requirements.txt
flask
flask-cors
pydantic
langchain
langchain-community
serpapi
python-dotenv

